requirements.txt
google-generativeai>=0.7.0
msal>=1.28.0
requests>=2.31.0
pandas>=2.2.0
python-dotenv>=1.0.0
faker>=24.0.0
